Added typing and changed tests
