+++ final_project/testing_pizzas.py
@@ -1,6 +1,4 @@
-""" Testing functions of Project """
-import re
-import terminal_commands as tc
+""" Testing functions of pizzas.py"""
 import pizzas
 
 
@@ -17,21 +15,3 @@ def test_dict_of_pizzas():
                                              'recipe': ['tomato sauce',
                                                         'mozzarella',
                                                         'tomatoes']}
-
-
-def test_deliver_of_tc():
-    """Test 3"""
-    result = tc.deliver(pizzas.Margherita('L'))
-    assert re.sub(r'[0-9]', '', result) == 'Margherita 🚴 Deliver for  min!'
-
-
-def test_pickup_of_tc():
-    """Test 4"""
-    result = tc.pickup(pizzas.Hawaiian('L'))
-    assert re.sub(r'[0-9]', '', result) == 'Hawaiian 🚴 Pick up for  min!'
-
-
-def test_preparing_of_tc():
-    """Test 5"""
-    result = tc.preparing(pizzas.Pepperoni('L'))
-    assert re.sub(r'[0-9]', '', result) == 'Pepperoni 🥘 Prepare for  min!'
